pyproject.toml
[project]
name = "evict-bot"
version = "0.1.0"
description = "Evict Discord Bot"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "asyncpg>=0.29.0",
    "humanfriendly>=10.0",
    "discord-ios>=1.0.0",
    "pomice>=2.9.0",
    "redis>=5.0.0",
    "aiohttp>=3.9.0",
    "jishaku>=2.5.0",
]

[tool.setuptools]
packages = ["bot", "cogs", "events", "rivalapi"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
